# priv/static/swagger.json
{
  "definitions": {
...
    "GroupUsers": {
      "description": "Relation between group and user",
      "example": {
        "group_id": 1,
        "program_date_of_joining": "2020/01/06",
        "program_manager_id": 1,
        "program_student_language": "English",
        "user_id": 1
      },
      "properties": {
        "group_id": {
          "description": "Id of a particular group",
          "type": "integer"
        },
        "program_date_of_joining": {
          "description": "Date of joining a program",
          "type": "utc_datetime"
        },
        "program_manager_id": {
          "description": "ID of a program manager",
          "type": "integer"
        },
        "program_student_language": {
          "description": "Language used in an enrolled program",
          "type": "string"
        },
        "user_id": {
          "description": "Id of a particular group",
          "type": "integer"
        }
      },
      "title": "Group User",
      "type": "object"
    },
...
    "School": {
      "description": "A school in the application",
      "example": {
        "block_code": "DOEAIDED",
        "board": "CBSE",
        "board_medium": "en",
        "category": "Government",
        "code": "872931",
        "district": "NORTH WEST DELHI",
        "district_code": "0701",
        "name": "Kendriya Vidyalaya - Rajori Garden",
        "region": "Urban",
        "state": "Delhi",
        "state_code": "DL",
        "type": "Open",
        "udise_code": "05040120901"
      },
      "properties": {
        "block_code": {
          "description": "Block Code",
          "type": "string"
        },
        "block_name": {
          "description": "Block Name",
          "type": "string"
        },
        "board": {
          "description": "Board",
          "type": "string"
        },
        "board_medium": {
          "description": "Medium",
          "type": "string"
        },
        "category": {
          "description": "Category",
          "type": "string"
        },
        "code": {
          "description": "Code",
          "type": "string"
        },
        "district": {
          "description": "District",
          "type": "string"
        },
        "district_code": {
          "description": "District Code",
          "type": "string"
        },
        "name": {
          "description": "Name",
          "type": "string"
        },
        "region": {
          "description": "Region",
          "type": "string"
        },
        "state": {
          "description": "State",
          "type": "string"
        },
        "state_code": {
          "description": "State Code",
          "type": "string"
        },
        "type": {
          "description": "Type",
          "type": "string"
        },
        "udise_code": {
          "description": "Udise Code",
          "type": "string"
        }
      },
      "title": "School",
      "type": "object"
    },
...
    "GroupIds": {
      "example": {
        "group_ids": [
          1,
          2
        ]
      },
      "properties": {
        "group_ids": {
          "description": "List of group ids",
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "Schools": {
      "description": "All schools in the application",
      "items": {
        "$ref": "#/definitions/School"
      },
      "title": "Schools",
      "type": "array"
    },
...
    "Sessions": {
      "description": "All the sessions",
      "items": {
        "$ref": "#/definitions/Session"
      },
      "title": "Sessions",
      "type": "array"
    },
...
    "Session": {
      "description": "A session in the application",
      "example": {
        "created_by_id": 1,
        "end_time": "2022-02-02T11:30:00Z",
        "form_schema_id": 1,
        "is_active": true,
        "meta_data": {
          "substitute-teacher-name": "Ms. Poonam"
        },
        "name": "Kendriya Vidyalaya - Weekly Maths class 7",
        "owner_id": 2,
        "platform": "meet",
        "platform_link": "https://meet.google.com/asl-skas-qwe",
        "portal_link": "https://links.af.org/kv-wmc7",
        "repeat_schedule": {
          "repeat-till-date": "2022-12-31 11:59:59",
          "repeat-type": "weekly"
        },
        "session_id": "c714-e1d4-5a42-0f9f-36b3",
        "start_time": "2022-02-02T11:00:00Z"
      },
      "properties": {
        "created_by_id": {
          "description": "User ID for the session creator",
          "type": "integer"
        },
        "end_time": {
          "description": "Session finish time",
          "type": "timestamp"
        },
        "form_schema_id": {
          "description": "Id for the form schema",
          "type": "string"
        },
        "is_active": {
          "description": "Tells whether session is active or not",
          "type": "boolean"
        },
        "meta_data": {
          "description": "Additional meta data for the session",
          "type": "map"
        },
        "name": {
          "description": "First name",
          "type": "string"
        },
        "owner_id": {
          "description": "User ID for the session owner",
          "type": "integer"
        },
        "platform": {
          "description": "Platform where session being hosted",
          "type": "string"
        },
        "platform_link": {
          "description": "Link for the platform",
          "type": "string"
        },
        "portal_link": {
          "description": "Link generated by the portal",
          "type": "text"
        },
        "repeat_schedule": {
          "description": "Repeat type and repeat till date for session",
          "type": "json"
        },
        "session_id": {
          "description": "Id for the session",
          "type": "string"
        },
        "start_time": {
          "description": "Session start time",
          "type": "timestamp"
        }
      },
      "title": "Session",
      "type": "object"
    }
  },
  "info": {
    "title": "DB Service application",
    "version": "1.0"
  },
  "paths": {
    "/api/session?session_id=c714-e1d4-5a42-0f9f-36b3": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.index",
        "parameters": [
          {
            "description": "The id the session",
            "in": "query",
            "name": "session_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "The name the session",
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          },
          {
            "description": "The platform id the session",
            "in": "query",
            "name": "platform_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sessions"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/session/{id}": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.show",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.update",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/school": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SchoolController.index",
        "parameters": [
          {
            "description": "The name the school",
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          },
          {
            "description": "The code the school",
            "in": "query",
            "name": "code",
            "required": false,
            "type": "string"
          },
          {
            "description": "The udise code the school",
            "in": "query",
            "name": "udise_code",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Schools"
            }
          }
        },
        "summary": "",
        "tags": [
          "School"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.SchoolController.create",
        "parameters": [
          {
            "description": "School to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        },
        "summary": "",
        "tags": [
          "School"
        ]
      }
    },
    "/api/session/{id}/update-groups": {
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.update_groups",
        "parameters": [
          {
            "description": "The id of the session record",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "List of group ids to update for the session",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupIds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/school/{schoolId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.SchoolController.delete",
        "parameters": [
          {
            "description": "The id of the school record",
            "in": "path",
            "name": "schoolId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "School"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.SchoolController.show",
        "parameters": [
          {
            "description": "The id of the school record",
            "in": "path",
            "name": "schoolId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        },
        "summary": "",
        "tags": [
          "School"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.SchoolController.update",
        "parameters": [
          {
            "description": "The id of the school record",
            "in": "path",
            "name": "schoolId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "School to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        },
        "summary": "",
        "tags": [
          "School"
        ]
      }
    },
...
    "/api/session": {
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.SessionController.create",
        "parameters": [
          {
            "description": "Session to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Session"
            }
          }
        },
        "summary": "",
        "tags": [
          "Session"
        ]
      }
    },
...
    "/api/group-user": {
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupUserController.index",
        "parameters": [
          {
            "description": "The id the group type",
            "in": "query",
            "name": "group_type_id",
            "required": false,
            "type": "string"
          },
          {
            "description": "The id the user",
            "in": "query",
            "name": "user_id",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupUser"
        ]
      },
      "post": {
        "description": "",
        "operationId": "DbserviceWeb.GroupUserController.create",
        "parameters": [
          {
            "description": "Group user to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupUser"
        ]
      }
    },
...
    "/api/group-user/{groupUserId}": {
      "delete": {
        "description": "",
        "operationId": "DbserviceWeb.GroupUserController.delete",
        "parameters": [
          {
            "description": "The id of the group user record",
            "in": "path",
            "name": "groupUserId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "",
        "tags": [
          "GroupUser"
        ]
      },
      "get": {
        "description": "",
        "operationId": "DbserviceWeb.GroupUserController.show",
        "parameters": [
          {
            "description": "The id of the group user record",
            "in": "path",
            "name": "groupUserId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupUser"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "DbserviceWeb.GroupUserController.update",
        "parameters": [
          {
            "description": "The id of the session user",
            "in": "path",
            "name": "groupUserId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Group user to create",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/GroupUsers"
            }
          }
        },
        "summary": "",
        "tags": [
          "GroupUser"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}